requirements.txt
Flask==2.3.2
line-bot-sdk==2.0.0
openpyxl==3.1.2
python-dotenv==1.0.0
